feat(chain): Replace protobuf with serde/BCS
The lack of bijective mapping and canonical serialisation in Protobuf
poses a challenge when attempted to be used for stable cryptographic
operations. While the tooling is pervasive and is widely adopted in the
ecosystem, better suited alternatives exist. Namely Borsh[^0] and BCS[^1],
with the latter being the choice for this project due to its reliance on
Serde[^2], which is the de facto standard for Rust serialisation needs.

[^0]: https://github.com/near/borsh
[^1]: https://github.com/diem/bcs
[^2]: https://serde.rs/

Signed-off-by: xla <self@xla.is>