feat(app): Complete ABCI interface
Completes the implementation of the ABCI interface from the app
perspective, which means it can run against a vanilla cometbft node with
version `v0.34.x`. All configured components within the app are
executed, that list is empty at the moment. An example implementation is
included by ways of a `Staking` skeleton, it lacks however any logic and
will crash the app if included in the component list.

Signed-off-by: xla <self@xla.is>