feat(node): Bootstrap node crate
Set of services on top of tower-abci[^0] to drive the interactions from
cometbft through ABCI. There is one service per connection[^1]
responsible to drive forward a worker. Every worker wraps an independent
instance of `App` for isolation guarantees and improved paralellism.

While it implements a run method which glues together all components,
this change-set does not add a binary which has a main, this is left for
a follow-up where it can be integrated as command in plz.

[^0]: https://crates.io/crates/tower-abci
[^1]: https://github.com/cometbft/cometbft/blob/main/spec/abci/abci%2B%2B_app_requirements.md#connection-state